requirements.txt
gin-config
torch
torchvision
timm
clip @ git+https://github.com/openai/CLIP.git@04f4dc2ca1ed0acc9893bd1a3b526a7e02c4bb10
Pillow
matplotlib
numpy
pytorch-lightning
